# backend/swagger.yaml
openapi: 3.0.0
info:
  title: MetLife Arena Operational API
  version: 1.0.0
  description: Real-time operational intelligence API for stadium details, concessions, and wait times.
servers:
  - url: http://localhost:3000/api
    description: Local development server
paths:
  /stadium:
    get:
      summary: Get full stadium database
      description: Returns the entire stadium document containing sectors, concessions, gate statuses, and emergency info.
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StadiumData'
        '401':
          description: Unauthorized (Missing or invalid API key)
        '500':
          description: Server Error (Database unavailable)
  /chat:
    post:
      summary: Send a message to the AI Stadium Assistant
      description: Sends a user message with location context and chat history to the Gemini-powered RAG assistant. Responses are grounded in the live stadium database.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: AI-generated response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Invalid input payload or prompt injection detected
        '429':
          description: Rate limit exceeded (100 requests per 15-minute window)
        '500':
          description: Gemini API error or internal server error
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    ChatRequest:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          maxLength: 500
          description: The user's chat message (max 500 characters)
        history:
          type: array
          maxItems: 30
          description: Previous conversation turns for multi-turn context
          items:
            type: object
            required:
              - role
              - text
            properties:
              role:
                type: string
                enum: [user, model]
              text:
                type: string
                maxLength: 2000
        currentLocation:
          type: string
          maxLength: 100
          description: The fan's current stadium sector/gate location
    ChatResponse:
      type: object
      properties:
        reply:
          type: string
          description: AI-generated response grounded in stadium data
        cached:
          type: boolean
          description: Whether this response was served from the spatial cache
    StadiumData:
      type: object
      properties:
        stadiumName:
          type: string
        sectors:
          type: array
          items:
            $ref: '#/components/schemas/Sector'
        gateStatus:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/GateStatus'
        emergencyInfo:
          type: object
          properties:
            emergencyNumber:
              type: string
            evacuationAssemblyPoints:
              type: string
            rules:
              type: array
              items:
                type: string
    Sector:
      type: object
      properties:
        id:
          type: string
        description:
          type: object
          properties:
            en:
              type: string
        gates:
          type: array
          items:
            type: string
        amenities:
          type: array
          items:
            $ref: '#/components/schemas/Amenity'
        concessions:
          type: array
          items:
            $ref: '#/components/schemas/Concession'
        restrooms:
          type: array
          items:
            $ref: '#/components/schemas/Restroom'
    GateStatus:
      type: object
      properties:
        status:
          type: string
        securityWaitMinutes:
          type: integer
          minimum: 0
        crowdDensity:
          type: string
    Amenity:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        location:
          type: string
        status:
          type: string
        details:
          type: string
    Concession:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        cuisine:
          type: string
        menu:
          type: array
          items:
            type: string
        vegetarianOptions:
          type: array
          items:
            type: string
        veganOptions:
          type: array
          items:
            type: string
        glutenFreeOptions:
          type: array
          items:
            type: string
        waitTimeMinutes:
          type: integer
          minimum: 0
        crowdStatus:
          type: string
    Restroom:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        types:
          type: array
          items:
            type: string
            enum: [Male, Female, Accessible, Family]
        waitTimeMinutes:
          type: integer
          minimum: 0
        crowdStatus:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
        details:
          type: string
          description: Additional context (only in non-production)